+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* Information about dataset versions can now be accessed via the the `.versions` property of a `Dataset` object. (#2)
+
 ## Version v0.1.0 - 2023-06-20
 
 Initial package release.